requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
Django==5.1.3
django-widget-tweaks==1.5.0
djhtml==3.0.7
pillow==11.0.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.2
urllib3==2.2.3
whitenoise==6.8.2
